had help with slight refactor of UpdateMovie component, fixed render on App component Route. UpdateMovie component functional
