+++ client/src/App.js
@@ -4,42 +4,30 @@ import SavedList from "./Movies/SavedList";
 import UpdateMovie from './Movies/UpdateMovie';
 import MovieList from "./Movies/MovieList";
 import Movie from "./Movies/Movie";
-import axios from 'axios';
 
 const App = () => {
   const [savedList, setSavedList] = useState([]);
-  const [movieList, setMovieList] = useState([]);
-
-  const getMovieList = () => {
-    axios
-      .get("http://localhost:5000/api/movies")
-      .then(res => setMovieList(res.data))
-      .catch(err => console.log(err.response));
-  };
 
   const addToSavedList = movie => {
     setSavedList([...savedList, movie]);
   };
 
-  useEffect(() => {
-    getMovieList();
-  }, []);
-
   return (
     <>
       <SavedList list={savedList} />
 
-      <Route exact path="/">
-        <MovieList movies={movieList} />
-      </Route>
+      <Route exact path="/" component={MovieList} />
 
-      <Route path="/movies/:id">
-        <Movie addToSavedList={addToSavedList} />
-      </Route>
+      <Route path="/movies/:id"
+        render={props => (
+        <Movie {...props} addToSavedList={addToSavedList} setSavedList={setSavedList}  /> )} 
+        />
 
-      <Route path="/update-movie/:id">
-        <UpdateMovie />
-      </Route>
+      <Route path="/update-movie/:id" 
+        render={props => (
+        <UpdateMovie {...props} savedList={SavedList} setSavedList={setSavedList} 
+        /> )} 
+        />
     </>
   );
 };
