+++ client/src/Movies/MovieList.js
@@ -1,12 +1,29 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import { Link } from "react-router-dom";
 import MovieCard from "./MovieCard";
+import axios from 'axios';
 
-function MovieList({ movies }) {
+
+function MovieList() {
+
+  const [movieList, setMovieList] = useState([]);
+
+  const getMovieList = () => {
+    axios
+      .get("http://localhost:5000/api/movies")
+      .then(res => setMovieList(res.data))
+      .catch(err => console.log(err.response));
+  };
+
+  useEffect(() => {
+    getMovieList();
+  }, []);
+
+
   return (
     <div className="movie-list">
       {
-        movies.map(movie => (
+        movieList.map(movie => (
           <Link key={movie.id} to={`/movies/${movie.id}`}>
             <MovieCard movie={movie} />
           </Link>
